Refactor PayPal controller into a controller and a service
This is to enable use of PayPal functionality from the Joiner controller
